requirements.txt
aiohttp==2.3.7
async-timeout==2.0.0
asyncio==3.4.3
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.7
idna==2.6
lxml==4.1.1
Markdown==2.6.10
multidict==3.3.2
mysqlclient==1.3.12
Pillow==5.0.0
pycodestyle==2.3.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.21.1
user-agent==0.1.9
yarl==0.17.0
